pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-contract-rag"
version = "0.1.0"
description = "English legal contract RAG with structure-aware chunking, reranking, and domain fine-tuning utilities."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
  "FlagEmbedding>=1.3.5,<2",
  "elasticsearch>=8.15,<9",
  "langchain>=0.3,<1",
  "langchain-elasticsearch>=0.3,<1",
  "langgraph>=0.3,<1",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "requests>=2.32,<3",
  "pyyaml>=6.0,<7",
  "scikit-learn>=1.5,<2",
  "tqdm>=4.66,<5",
  "torch>=2.3,<3",
  "transformers>=4.44,<5",
  "sentence-transformers>=3.0,<4",
]

[project.optional-dependencies]
dev = ["pytest>=8.2,<9", "pytest-cov>=5.0,<6"]

[project.scripts]
legal-rag-prepare-corpus = "legal_contract_rag.cli:prepare_corpus_main"
legal-rag-build-training = "legal_contract_rag.cli:build_training_main"
legal-rag-run-eval = "legal_contract_rag.cli:run_eval_main"
legal-rag-prepare-flagembedding = "legal_contract_rag.cli:prepare_flagembedding_data_main"
legal-rag-train-embedder = "legal_contract_rag.cli:run_flagembedding_embedder_main"
legal-rag-train-reranker = "legal_contract_rag.cli:run_flagembedding_reranker_main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
